fix: make stack navigator work with latest gesture handler (#10270)
With the release of Gesture Handler 2 some gesture handlers have been updated to follow state-flow more consistently across the platforms. Among them was the Pan gesture which was changed to transition to the `BEGAN` state as soon as finger touches the screen opposed to just before transitioning to the `ACTIVE` state.
`@react-navigation/stack` is relying on the old behavior which may be causing some unexpected issues in apps, for example: https://github.com/software-mansion/react-native-gesture-handler/issues/1801.

This PR changes the swipe-back gesture to start when the Pan transitions to the`ACTIVE` state. Since previously Pan gesture was transitioning to the `BEGAN` state just before transitioning to the `ACTIVE` state this change would be compatible with older versions of Gesture Handler.